custom block size, int quantization

+++ project/processing/quantization.h
@@ -5,10 +5,10 @@
 
 class Quant {
 public:
-    static Matrix<double> applyQuantization(const Matrix<double> &matrix, size_t size_of_block, double q);
+    static Matrix<int> applyQuantization(const Matrix<double> &matrix, size_t size_of_block, double q);
 
 private:
-    static Matrix<double> getQMatrix(size_t N, double q);
+    static Matrix<int> getQMatrix(size_t N, double q);
 
 };
